shift-backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shift-backend"
version = "0.1.0"
description = "FastAPI backend with async PostgreSQL and Alembic migrations."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.32,<3.0.0",
  "asyncpg>=0.29.0,<1.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "alembic>=1.13.2,<2.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "google-auth>=2.35.0,<3.0.0",
  "requests>=2.32.0,<3.0.0",
  "pymysql>=1.1.1,<2.0.0",
  "pyodbc>=5.1.0,<6.0.0",
  "oracledb>=2.2.1,<3.0.0",
  "snowflake-connector-python>=3.12.0,<4.0.0",
  "firebird-driver>=1.10.6,<2.0.0",
  "fdb>=2.0.4,<3.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
